+++ phandas/__init__.py
@@ -3,7 +3,7 @@ A multi-factor quantitative trading framework for cryptocurrency markets.
 """
 
 __author__ = "Phantom Management"
-__version__ = "0.12.4"
+__version__ = "0.12.5"
 
 from .core import Factor
 from .panel import Panel
